requirements.txt
discord.py>=1.3.0
dblpy>=0.3.3
python-Levenshtein>=0.12.0
aiohttp>=3.6.2
asyncpg>=0.20.1
python-dotenv>=0.13.0
yoyo-migrations>=7.0.2
psycopg2>=2.8.5
